+++ src/prisma.service.ts
@@ -1,9 +1,29 @@
-import { Injectable, OnModuleInit } from '@nestjs/common';
-import { PrismaClient } from '@prisma/client';
+import 'dotenv/config';
+import { Injectable, OnModuleInit, OnModuleDestroy } from '@nestjs/common';
+import { PrismaClient } from './generated/prisma/client';
+import { PrismaPg } from '@prisma/adapter-pg';
 
 @Injectable()
-export class PrismaService extends PrismaClient implements OnModuleInit {
+export class PrismaService extends PrismaClient implements OnModuleInit, OnModuleDestroy {
+  constructor() {
+    const connectionString = process.env.DATABASE_URL;
+    const schema = new URL(connectionString).searchParams.get('schema') ?? undefined;
+    const adapter = new PrismaPg(
+      {
+        connectionString: connectionString,
+        options: schema && `-c search_path="${schema}"`,
+      },
+      { schema }
+    );
+
+    super({ adapter });
+  }
+
   async onModuleInit() {
     await this.$connect();
+  }
+
+  async onModuleDestroy() {
+    await this.$disconnect();
   }
 }