+++ src/utils/classes/PDFGenerator.ts
@@ -1,4 +1,4 @@
-import * as PDFDocument from 'pdfkit';
+import PDFDocument from 'pdfkit';
 
 export default class PDFGenerator {
   doc = null;